services/visualization_consumer/requirements.txt
confluent_kafka
watchdog==2.1.9
Pillow
streamlit